lxd: Move instance name validator to instancetype package
