+++ lxd/instance/drivers/driver_qemu.go
@@ -5168,7 +5168,7 @@ func (d *qemu) Rename(newName string, applyTemplateTrigger bool) error {
 	d.logger.Info("Renaming instance", ctxMap)
 
 	// Quick checks.
-	err = instance.ValidName(newName, d.IsSnapshot())
+	err = instancetype.ValidName(newName, d.IsSnapshot())
 	if err != nil {
 		return err
 	}
